+++ pages/Main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import BusSchedules from './BusSchedules';
 import BusRoute from './BusRoute';
-import Contacts from './Contacts';
+import Contacts from './EmptyContacts';
 
 
 import MainStyles from '../styles/MainStyles';
